==> manualTrainingSetSelection/07_forTrainingPrep.ipynb <==
# This file combines the no change (good) and no change (bad) into one feature no change. 
##### This allows us to create a binary classification model and simplify our analysis into weather a location has changed or not.

import pandas as pd

df = pd.read_csv('./csv/imgPairsForTraining.csv')

change = list(df.change.dropna())
change_name = [1]*len(change)
good = list(df.good.dropna())
bad = list(df.bad.dropna())
comb = good+bad
no_change_name = [0]*(len(comb))
# 1 if change, 0 if no change

new_df = pd.DataFrame(data={"img":change+comb,"change":change_name+no_change_name})
new_df = new_df.astype("Int64")

new_df.to_csv('./csv/imgsForTrainingByChange.csv',index=False)
